[Refactor]: LottoNo -> LottoNumber 로 클래스명 변경
